+++ changelog.md
@@ -1,6 +1,7 @@
 # v1.0.1
 
 - fish spinner available at the mod settings :fire:
+- Added custom entering and exiting editor loading texts
 
 # v1.0.0